+++ src/com/createsend/Segments.java
@@ -21,19 +21,21 @@
  */
 package com.createsend;
 
-import java.util.Date;
-
-import javax.ws.rs.core.MultivaluedMap;
-
 import com.createsend.models.PagedResult;
-import com.createsend.models.segments.*;
-import com.createsend.models.subscribers.Subscriber;
+import com.createsend.models.segments.ClauseResults;
+import com.createsend.models.segments.RuleCreationFailureDetails;
+import com.createsend.models.segments.RuleGroup;
+import com.createsend.models.segments.Segment;
+import com.createsend.models.subscribers.SubscriberWithJoinedDate;
 import com.createsend.util.AuthenticationDetails;
 import com.createsend.util.ErrorDeserialiser;
 import com.createsend.util.JerseyClientImpl;
 import com.createsend.util.exceptions.CreateSendException;
 import com.createsend.util.jersey.JsonProvider;
 import com.sun.jersey.core.util.MultivaluedMapImpl;
+
+import javax.ws.rs.core.MultivaluedMap;
+import java.util.Date;
 
 /**
  * Provides methods for accessing all <a href="http://www.campaignmonitor.com/api/segments/" target="_blank">
@@ -141,7 +143,7 @@ public class Segments extends CreateSendBase {
      * @see <a href="http://www.campaignmonitor.com/api/segments/#getting-active-subscribers" target="_blank">
      * Getting active subscribers</a>
      */
-    public PagedResult<Subscriber> active() throws CreateSendException {
+    public PagedResult<SubscriberWithJoinedDate> active() throws CreateSendException {
     	return active(1, 1000, "email", "asc", false);
     }
 
@@ -153,7 +155,7 @@ public class Segments extends CreateSendBase {
      * @see <a href="http://www.campaignmonitor.com/api/segments/#getting-active-subscribers" target="_blank">
      * Getting active subscribers</a>
      */
-    public PagedResult<Subscriber> active(boolean includeTrackingPreference) throws CreateSendException {
+    public PagedResult<SubscriberWithJoinedDate> active(boolean includeTrackingPreference) throws CreateSendException {
         return active(1, 1000, "email", "asc", includeTrackingPreference);
     }
 
@@ -168,7 +170,7 @@ public class Segments extends CreateSendBase {
      * @see <a href="http://www.campaignmonitor.com/api/segments/#getting-active-subscribers" target="_blank">
      * Getting active subscribers</a>
      */
-    public PagedResult<Subscriber> active(Integer page, Integer pageSize,
+    public PagedResult<SubscriberWithJoinedDate> active(Integer page, Integer pageSize,
         String orderField, String orderDirection) throws CreateSendException {
     	return active("", page, pageSize, orderField, orderDirection, false);
     }
@@ -185,7 +187,7 @@ public class Segments extends CreateSendBase {
      * @see <a href="http://www.campaignmonitor.com/api/segments/#getting-active-subscribers" target="_blank">
      * Getting active subscribers</a>
      */
-    public PagedResult<Subscriber> active(Integer page, Integer pageSize, String orderField,
+    public PagedResult<SubscriberWithJoinedDate> active(Integer page, Integer pageSize, String orderField,
         String orderDirection, boolean includeTrackingPreference) throws CreateSendException {
         return active("", page, pageSize, orderField, orderDirection, includeTrackingPreference);
     }
@@ -204,7 +206,7 @@ public class Segments extends CreateSendBase {
      * @see <a href="http://www.campaignmonitor.com/api/segments/#getting-active-subscribers" target="_blank">
      * Getting active subscribers</a>
      */
-    public PagedResult<Subscriber> active(Date subscribedFrom, Integer page, Integer pageSize,
+    public PagedResult<SubscriberWithJoinedDate> active(Date subscribedFrom, Integer page, Integer pageSize,
         String orderField, String orderDirection) throws CreateSendException {
     	return active(JsonProvider.ApiDateFormat.format(subscribedFrom),
     			page, pageSize, orderField, orderDirection, false);
@@ -225,13 +227,13 @@ public class Segments extends CreateSendBase {
      * @see <a href="http://www.campaignmonitor.com/api/segments/#getting-active-subscribers" target="_blank">
      * Getting active subscribers</a>
      */
-    public PagedResult<Subscriber> active(Date subscribedFrom, Integer page, Integer pageSize,
+    public PagedResult<SubscriberWithJoinedDate> active(Date subscribedFrom, Integer page, Integer pageSize,
         String orderField, String orderDirection, boolean includeTrackingPreference) throws CreateSendException {
         return active(JsonProvider.ApiDateFormat.format(subscribedFrom),
             page, pageSize, orderField, orderDirection, includeTrackingPreference);
     }
 
-    private PagedResult<Subscriber> active(String subscribedFrom, Integer page, Integer pageSize,
+    private PagedResult<SubscriberWithJoinedDate> active(String subscribedFrom, Integer page, Integer pageSize,
         String orderField, String orderDirection, boolean includeTrackingPreference) throws CreateSendException {
         MultivaluedMap<String, String> queryString = new MultivaluedMapImpl(); 
         queryString.add("date", subscribedFrom);